feat: breakout BaseProtocol
